Implement 'add project' button functionality
This commit implements the functionality for the 'add project' button in
the sidebar. When the button is clicked, a form appears and the user is
prompted to input the name of the project and whether to cancel or confirm
and add the project. Currently it only requires the name of the project,
but more requirements and fields may be added.